Refactor biome generators into separate files
