+++ server/src/physics/entity.rs
@@ -35,7 +35,6 @@ impl<'a> System<'a> for EntityPhysicsSystem {
             chunk_map,
             entities,
         ) = data;
-
         // Go through entities and update their positions according
         // to their velocities.
 
@@ -56,6 +55,7 @@ impl<'a> System<'a> for EntityPhysicsSystem {
         )
             .join()
         {
+            debug!("iter");
             let mut velocity = restrict_velocity.get_unchecked().clone();
 
             let mut pending_position = position.current + velocity.0;
