+++ server/src/worldgen/biomes/mod.rs
@@ -1,0 +1,7 @@
+//! Biome grid creation.
+
+mod distorted_voronoi;
+mod two_level;
+
+pub use distorted_voronoi::DistortedVoronoiBiomeGenerator;
+pub use two_level::TwoLevelBiomeGenerator;
